When factory didn't return a model, return a simple scoped model.
Usually factory decides what to create based on the field in document. So in order for the factory to properly execute, it needs to have this document already loaded into the client model. To work around those chicken/egg issues, we need to return a simple scope model if the document is not in the model yet. And after it was fetched, the user needs to recreate the scope model by doing `myModel = myModel.at()` to trigger the factory once again.